tests/wpt/tests/css/css-borders/tentative/parsing/corner-shape-valid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'corner-shape' with valid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shaping">
<meta name="assert" content="This test checks that 'corner-shape' supports the values 'round' and 'angle'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_valid_value("corner-shape", "round");
test_valid_value("corner-shape", "angle");
test_valid_value("corner-shape", "angle round");
test_valid_value("corner-shape", "round angle");
test_valid_value("corner-shape", "angle round angle");
test_valid_value("corner-shape", "angle round angle round");
</script>
